refactor: connect RsvpForm to useRsvpFormFields hook
