+++ src/components/rsvp/RsvpForm.tsx
@@ -1,4 +1,4 @@
-import type { RsvpItem, RsvpStatus } from "../../types/rsvp";
+import type { RsvpStatus } from "../../types/rsvp";
 
 type Props = {
   guestName: string;
@@ -10,7 +10,9 @@ type Props = {
   status: RsvpStatus;
   setStatus: (v: RsvpStatus) => void;
 
-  onAdd: (item: RsvpItem) => void;
+  resetForm: () => void;
+
+  onAdd: () => void;
 };
 
 export default function RsvpForm({
@@ -20,6 +22,7 @@ export default function RsvpForm({
   setEmail,
   status,
   setStatus,
+  resetForm,
   onAdd,
 }: Props) {
   function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
@@ -28,22 +31,11 @@ export default function RsvpForm({
     const name = guestName.trim();
     const mail = email.trim();
 
+    // Basic UI validation (presentation concern)
     if (!name || !mail) return;
 
-    const item: RsvpItem = {
-      id: crypto.randomUUID(),
-      guestName: name,
-      email: mail.toLowerCase(),
-      status,
-      createdAt: new Date().toISOString(),
-    };
-
-    onAdd(item);
-
-    // reset
-    setGuestName("");
-    setEmail("");
-    setStatus("Going");
+    onAdd();
+    resetForm();
   }
 
   return (
